+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.4.1] - 2025-06-30
+
+## Added
+
+- Internal observable extensions class.
+
+### Changed
+
+- PrivMX session script now tracks lib connection.
+
 ## [2.3.4] - 2025-06-25
 
 ### Changed